refactor: Wrap scheme closure with class Closure
