+++ src/system/interpreter.js
@@ -1,9 +1,10 @@
 import * as _ from "../deps/underscore-1.10.2-esm.js"
 import { TopEnv, CoreEnv, nil, undef, max_trace_size } from "../header.js"
-import { isSymbol, isClosure, makeClosure } from "./_types.js"
+import { isSymbol } from "./_types.js"
 import { to_write, inspect } from "./_writer.js"
 import Call from "./call.js"
 import Class from "./class.js"
+import { Closure, isClosure } from "./closure.js"
 import Compiler from "./compiler.js"
 import { BiwaError, Bug } from "./error.js"
 import { Pair, List } from "./pair.js"
@@ -142,11 +143,8 @@ const Interpreter = Class.create({
       v[i+1] = this.index(s, i);
     v[n+1] = dotpos;
 
-    if(dotpos == -1)
-      v.expected_args = args;
-
-    makeClosure(v);
-    return v;
+    const expected_args = dotpos == -1 ? args : undefined;
+    return new Closure(v, expected_args);
   },
 
   // private
@@ -176,12 +174,16 @@ const Interpreter = Class.create({
   },
 
   // private
+  // a: arbitary object (temporary register)
+  // x: opecode
+  // f: integer
+  // c: BiwaScheme.Closure
+  // s: integer
   _execute: function(a, x, f, c, s){
     var ret = null;
     //Console.puts("executing "+x[0]);
     
     while(true){ //x[0] != "halt"){
-
       this.run_dump_hook(a, x, f, c, s);
 
       switch(x[0]){
@@ -194,7 +196,7 @@ const Interpreter = Class.create({
         break;
       case "refer-free":
         var n=x[1], x=x[2];
-        a = c[n+1];
+        a = c.il[n+1];
         this.last_refer = "(anon)";
         break;
       case "refer-global":
@@ -249,7 +251,7 @@ const Interpreter = Class.create({
         break;
       case "assign-free":
         var n=x[1], x=x[2];
-        var box = c[n+1];
+        var box = c.il[n+1];
         box[0] = a;
         a = undef;
         break;
@@ -309,10 +311,11 @@ const Interpreter = Class.create({
         var n_args = this.index(s, 0);
         if(isClosure(func)){
           a = func;
-          x = func[0];
+          const op = func.il;
+          x = op[0];
 
           // The position of dot in the parameter list.
-          var dotpos = func[func.length-1];
+          var dotpos = op[op.length-1];
 
           if (dotpos >= 0) {
             // The dot is found
@@ -348,7 +351,7 @@ const Interpreter = Class.create({
             }
           }
           f = s;
-          c = a;
+          c = func;
         }
         else if(func instanceof Function){ // Apply JavaScript function
           // load arguments from stack
